requirements.txt
spacy[apple]
spacy-langdetect
PyPDF2
python-docx
streamlit
langdetect